+++ docs/tr1/CHANGELOG.md
@@ -30,6 +30,7 @@
     - added the ability for stacks of movable blocks to fall and land on trapdoors, drawbridges, bridges, sliding pillars, and falling blocks
     - added the ability for stacks of movable blocks to fall when on opened trapdoors and drawbridges
     - fixed various bugs with falling movable blocks
+- fixed pushblocks becoming unusable when on the same sector as a door that does not sit on a room portal (#3814)
 - fixed recordings replaying commands twice (regression from 4.14)
 - fixed the fix for the sticky corner glitch not being optional - now linked to Gameplay → Fixes → Wall glitch mode (#3957, regression from 4.14)